moved initializeGardenInformation to CreateGardenActivity
